Split GitHub Projects parser and transport layers
